Remove automatic naming of tmux panes when using ssh
(1) If I press ctrl-c during the connection, it keeps the name, which is confusing
(2) When I use the shortcuts, like 'o' for Oxford, it's hard to understand
(3) If I rename the pane while connected, it will be changed back afterwards

All of these are fixable with some additional logic, but I'm not convinced it's useful enough.
Especially as I often end up with multiple connections to each server.